# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "StoreFy API",
        "description": "StoreFY Swagger OpenApi description",
        "contact": {
            "email": "georgeton.ufg@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Api local server"
        },
        {
            "url": "http://storefy.workingtech.com.br:31080/api",
            "description": "Api dev server"
        }
    ],
    "paths": {
...
        "/categorias": {
            "get": {
                "tags": [
                    "categorias"
                ],
                "summary": "Get all items",
                "operationId": "getCategorias",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                },
                                                "tipo": {
                                                    "title": "tipo",
                                                    "description": "Retorno do tipo de categoria"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/Categorias"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "categorias"
                ],
                "summary": "Add a new item",
                "operationId": "addCategorias",
                "requestBody": {
                    "description": "Object to add new item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categorias"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/Categorias"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/validate_message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categorias/{uuid}": {
            "get": {
                "tags": [
                    "categorias"
                ],
                "summary": "Get item by ID",
                "operationId": "getCategoriasById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                },
                                                "tipo": {
                                                    "title": "tipo",
                                                    "description": "Retorno do tipo de categoria"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/Categorias"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "categorias"
                ],
                "summary": "Update item",
                "operationId": "updateCategorias",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Object to update item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categorias"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/Categorias"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "categorias"
                ],
                "summary": "Delete item",
                "operationId": "deleteCategorias",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "validate_message": {
                "title": "Error message",
                "description": "Error jso1n",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "readOnly": true
                    },
                    "message": {
                        "type": "string",
                        "minLength": 8,
                        "readOnly": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ErrorMessage"
                }
            },
...
            "Categorias": {
                "title": "Categorias model",
                "description": "Categorias model",
                "required": [],
                "properties": {
                    "nome": {
                        "title": "nome",
                        "description": "Nome"
                    },
                    "uuid": {
                        "title": "uuid",
                        "description": "Uuid"
                    },
                    "tipo": {
                        "title": "tipo",
                        "description": "Tipo de categoria (Ex: cat_clientes, cat_contas)"
                    },
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "default": false
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Categorias"
                }
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}